better support for generated files and "generator plugins"
Support generating files using a similar syntax to normal compilation, e.g.:

sed-y = a.c
a.c-y = a.c.in
a.c-SED_SCRIPT-y := 's,a,b,'

This will generate a.c (if some other target has it as a prerequisite).

The recipe and other rules can be provided by so-called generator plugins.
That is, makefiles that define a $x_rule and $x_recipe macro, that are included
before kmake.mk (kmake.mk will evaluate those if $x appears in extra-gen).

Also, $x-suffix can be specified to implement finding implicit source files.
For example, for sed-suffix := c.sed, then the above a.c will be generated
from a.c.sed if a.c-y empty.