fix install-strip
When installing with libtool, then libtool only recognizes the -s flag, and
does magic for shared libraries (runs strip separately). For programs, -s is
forwarded to install. --strip-program is forwarded always.

libtool doesn't recognize --strip-program, so install sees that option
for shared libraries without getting -s as well, which leads to an error.
Fortunately, the local libtool (intalled via libtoolize) seems to
run the correct strip program so we don't need that option for shared libs.

Solution is to pass -s --strip-program=strip only for programs,
while for shared libraries only -s is passed.

Also add a initial value for STRIP which was missing yet, and simplify
the install-lib-% recipe.