improved sub-makefile support
Invoking "$(MAKE) submake-all" requires knowing the name of the stub makefile
that ought to include kmake.mk. This is fragile should be avoided, and invoking
make is bad for performance anyway.

Instead, set up phony targets and depenencies such that local targets are
updated first, and following that comes the invocation of the subdirectory
make.

It's an explicit choice that sub-makefiles are executed only after all kmake
targets are finished because kmake doesn't have any dependency information, and
sub-makefile targets might depend on kmake targets, so the sub-makefile targets
cannot be easily built in parallel to kmake targets.