.cmd file rework
- old solution was slow despite the optimization efforts using a hash
- apparently there were too many subprocesses involved, just for avoiding
  to read the entire .cmd file, but that seems to be slower actually
- the .cmd file size is usualy only 1 or 2 pages which will be
  read completely by the OS anyway

The new solution works a bit like the depfiles (.dep). For each
object file a make fragment is generated, and included in subsequent
make calls. The fragment contains a small rule:

$object: OLDCMD = $(cmdline)

This way a target-specific variable is created, containing the
previous command line. This is used to check if the command
line changed, and if yes the .cmd is simply touched, causing
the object to be rebuild.