add cmd files for generators to cleanfiles
Previously there were simply not cleaned, but it became more of an issue
because now .oldcmd is included but only if it's mentioned in cleanfiles.

So this fixes the .cmd file feature for generated files.

The files will end up in cleanfiles even if the generator does not
use the cmd file feature. But this is not a problem, this the .cmd
file won't be created (no dependency on it), and as a result the
.oldcmd will also not exist (and there is no rule to created it). So
it doesn't matter, there are just some superflous calls to rm -f.