objects in the top directory erroneously had a ./ prefix
kmake generally avoids the ./ prefix for all files, and references files by
their name. For files in the top directory this is just the filename without
directory part. Wrong use of $(dir) may introduce that prefix by accident.

Simplify the relevant function prefixtarget. It's never called with a list of
files so it doesn't need to handle those. Also clarify some comments.