ensure .oldcmd is craeted and yields a non-empty OLDCMD
If a generator uses the .cmd file feature but provides an empty CMD. then the
.oldcmd file contained an empty OLDCMD variable. This is equivalent to a
non-existent .oldcmd. If CMD continued to be empty the .cmd wouldn't be
created. This caused make to update targets that depend on it (special behavior
if a recipe for an ordinary file does not create it).

This is solved by enforcing that the .oldcmd file assignts at least one
character to OLDCMD, causing the .cmd file to be created initiallly.