new property compiler
Allows to select the compiler program. In practice, this can be
used to enforce gcc even for C++ code (to avoid linking libstd++).
Or to use clang instead of gcc.

As with properties, compiler can be set on a any vars or targets, and
targets inherit from vars.

Also generators may add more properties.