place KM_CHECKFLAGS after the test program
The idea was that the driver is an interpreter that runs the test program and that
arguments via KM_CHECKFLAGS get forwarded to the test program, but the current
implementation makes that hard.

Arguments to the actual driver can be harded coded in a wrapper script that becomes the
driver.