new vars noinstprogs and noinstlibs
These can be used to avoid installing the affected target, while still building
it unconditionally, which is especially useful for libtool convinience
libraries or to build programs that should not be installed (compile / link
checkers).

Internally this is implemented by a special handling of the "noinst" value of
the dir property. Actual values should be absolute paths (though this is not
enforced yet) so there is no ambigiuty.