add headers from $var-y as order-only dependency to objects
In case headers are generated they must be added as a dependency somewhere,
e.g. by adding them to $var-y. The previous commit allows header in $var-y,
but that only forms a dependency for programs and libraries to the headers.

But individual object files are the actual items that need the dependency
(maybe, knowing if an object really depends on a header requires compiling it).
The aformentioned dependency still allows for race conditions since object
files compilation and header generation may happen in parallel.

This change adds a dependency for object files to headers, but uses the
order-only style since which  does what we want:
- If an object does not exist, ensure the generated header exists, it may
  depend on the header
- If an object exists, the order-only dependency is irrelevant. Instead,
  the truth is know available in the .dep file.
- If the header is not actually generated but an ordinary file, the
  order-only style doesn't add extraneous update-if-changed dependencies.