built-in "generated" pseudo-generator
Sometimes you want to generate files that don't fit into bin-y and friends.
In that case they can be listed in the generated-y list.

This is similar to the byproduct generator but doesn't add any dependencies.
Everything is assumed to be provided by custom rules.

$(objdir)file:
    # to stuff

Marking such files as generated has some benefits, for example vpath is
set correctly for out-of-tree builds or special INCLUDE flag handling.

Another typical use case is when libraries are produced by sub-make calls.
We don't otherwise know what files are produced by a sub-make. But if a
kmake-produced file depends on a sub-make file, then such a dependency can
be declared with the help of this pseudo-generator (not pretty but it works).

E.g.

submake-y := foo/
generated-y := foo/libfoo.la

$(objdir)libfoo.la: submake-all-$(srcdir)foo/

mytool-y := tool.c foo/libfoo.la