basic support for package distributing (tarball)
kmake can create tarballs for the package, providing some heuristics
what is to be distributed and what not.

New targets are dist, dist-xz, dist-gz and dist-bz2. The former creates
the tarball distrubtion in a subdirecotry (see DISTDIR notes below) and the
other three create a compressed tarball from that directory.

By default, all files that are mentioned in a var's list ($var-y), unless
it can be determined to be a generated file. This includes files listed in the
built-in vars headers, data and sysconf. Also, kmake adds its own files to
the distribtion, but not the wrapper file that kmake.mk is included from!

More files can be distributed through the dist var (dist-y := some_file). When
the heuristic fails, then files can be excluded throught the nodist var, for
example if a file is generated by an external facility (e.g. configure script).

The subdirectory can be controlled via DISTDIR and D on the make command line.
DISTDIR must be an absolute path, while D must be relative to the source
tree, e.g.

make dist D=mydir