new var headers
Just like data and sysconf except files are installed into $(includedir)
by default. Of course, the use case is to install development headers.

This may also be needed to get generated headers generated in the first place
if there is no other target requiring this because data files are built by
default.