Revert "replace file list with a pattern in standard recipes"
The pattern method failed when a file a.c included a another file b.c. Due
to the generated dependency file b.c ends up in prerequisite list and it
matches the pattern, so kmake would attempt to generate one object file
from multiple .c files.

This is bad, so the original problem is solved with a more robust
solution coming next.

This reverts commit d52fb00422e613c7a99cda30f60d024a7de7c174.