make use of automatic variables again
automatic variables are necessary since they are fixed-up for vpath
by make. The PARTS variable was not and broke out-of-tree builds because
the linker was passed wrong paths to object files.

Now, automatic variables are used again. However, in order to not
re-introduce excess file lists the automatic variables are filtered
by the PARTS variable so unexpected files are not used.

Fixes: e4dd39c9 don't use automatic variables in builtin kmake rules