properly handle targets that are in a subdirectory
If you say bin-y := path/to/prog, then this is properly handled now.
Previously "path/to" was ignored, leading to a flawed built.

The name of the subdir.mk that defines the target is now recorded, to tell the
paths apart later on (path/to/subdir.mk vs path/to/path/to/prog). Maybe this
can be useful for diagnostics as well in the future.

This change also sanitises if source files are given with leading ./. This
must be stripped to not confuse vpath magic down the road.