completely rework install goal
Previously, install targets were phony targets that installed a "static"
file list (in a squential fashion). This had a number of disadvantages

- the file list was in discovery order, as subdir.mk are processed
- libtool isn't smart enough if it gets a list of files but
  not in the correct order (relinking on installation would fail)
- the sequential method is also really slow
- the file list wasn't subject to vpath magic
  libs and programs assumed to be in $(OUTDIR), data assumed to be in $(SRCDIR)
  -> could not install generated data

Now, the install targets uses a normal recipe, with the to-be-installed file
as first prerequisite. Other prequisite may include libraries that the installed
file depends on, so that lower-level libraries are installed first.

This has the advantange that installation is fully parallized (if
library dependencies permit) and the installed file is subject to vpath,
which does the right thing.

The implication is that make install also always builds all files to be
installed, whereas previous make install would not build anything.