add a test case for $$(srcdir) behavior of sub-directory flags
Using $$(srcdir) in subdir-CFLAGS-y should evaulate to the directory
where the flag value is effective, i.e. not the directory where subdir-CFLAGS-y
is defined (as opposed to just $(srcdir)). If there are mutiple levels of
sub-directories below the subdir.mk that defines subdir-CFLAGS-y, when each
level gets its appropriate value for $$(srcdir).