flock locking between link and install of .la files
Since the install rework files are installed concurrently. This is
generally a good thing. But there is an issue with that leads to intermittent
failures, when it links a library and one of the dependency libraries
is insalled at the same time.

Regulating access to .la files with flock locking seems to help immensely.
See comment for additional details.