improve handling of makefile fragments
When make includes a fragment, it searches for a recipe for it, perhaps
runs it and re-executes itself if it finds the fragment was updated.

This dance is unecessary for .dep and .oldcmd files and is costly, because
both types are created implicitely, so make didn't have a recipe yet, so
it always re-executed itself if it found one of the files to be updated
(and .oldcmd is always updated). Luckily it can be avoided, by
specifiing empty recipes for .dep and .oldcmd.

This is based on the findings found at
http://make.mad-scientist.net/papers/advanced-auto-dependency-generation/.
It also suggests to not use implicit rules for these kind of files;
follow that suggestion.

Then there was also an issue that .oldcmd was always created but not the
corresponding .cmd under make -q. The .oldcmd files are created
via make functions, they are always executed. .cmd is created by a touch command
which is not run under make -q. Change the logic a bit to also create
.oldcmd through a command (renamed from a temporary file).