automatically add order-only dependencies for generated files via INCLUDES
When a target uses the INCLUDES flag, and we know about generated files
in any of the directories listed there, then we can add an order-only
dependency on that generated files. Similarly to how we automatically
add -I$(OUTDIR)path/to/dir for those directories.

This helps to avoid to manually specify dependencies on generated headers
(which is still possible by listing them in $var-y).

The output directory of a target is treated the same way (if it contains
generated files).