emit vpath rules for all progs
Since commit 7007f45bc9 (do not emit build rules for targets that
aren't actually built) vpath was also avoided for targets that
aren't build by kmake.

This breaks targets with custom rules, e.g.
```
$(objdir)myprog:
	cp […] $@
```

This should be supported and needs vpath to be set correctly for myprog,
if O=... is used.