try to protect a bit against misuse of variables in recipes
clear out some variables and set srcdir/objdir to bogus values after
processing all subdir.mk files. All those variables must not be used
in recipes because they would contain the values of the last subdir.mk
that was processed.

If a subdir.mk defines a custom recipe (which in itself is fine), it must not
use variables that change throughout processing the other subdir.mk files.

srcdir and objdir get values that clearly indicate the situation.