flock locking between link and install of .la files, part 2
It turns out the flock trick needs to be expanded to the dependencies
of the affected libraries as well, i.e. work recursively.

Additionally, the list of locked libraries is based on the target-specific
PARTS variable because that which does not include $(OUTDIR). This is needed
the output of the (newly added) getdeps_r function never includes $(OUTDIR)
(being the recursive variant of getdeps) so the flock command generator needs
to add $(OUTDIR) to all files. With $^, $(OUTDIR) prefix would be doubled.