handle subdir-XXX-y flags if already set by the master Makefile
If the master Makefile populates subdir-y, then it makes sense
to also populate subdir-XXX-y flags there, to avoid a top-level subdir.mk.

Also, tighten subdir-y handling a bit. Either set it in the master Makefile
or in a top-level subdir.mk, but not both (the initial subdir-y may be empty / .
OR a list of directories but not a mix of . and directories).

Add a guard for . in subdir-y set in a subdir.mk, to prevent accidental recursion.