fix hooks for top-level subdir.mk
all_subdirs did not contain an entry for the top directory
if it contains the root subdir.mk, because $(srcdir) was
empty in this case.

To fix, build up all_subdirs and hooks directly in inc_subdir(),
before $(srcdir) is made empty.

process-subdir.mk still sees a empty $(srcdir), and that's presented
to the root subdir.mk, no change in this regard.