basic support for building shared objects without libtool
Syntax as usual:

libs-y := libfoo.so
libfoo.so-y := foo.c

libfoo.so is installed to $(libdir).

At this point, basically only building the shared object is supported but not
linking against it. Because when linking using `foo-DEPS-y := bar/libbaz.so`
then `bar/libbaz.so` will be passed straight to the linker. And the linker does
not strip the path so foo would have a dependency to "bar/libbaz.so" which
cannot be statisfied after installing (libbaz.so is in $(libdir) without bar
prefix). Maybe I'll fix that later, for now libtool is clearly recommended for
any advanced stuff.