make ensure_slash work with input "/"
ensure_slash removed the slash at the end first if there was one
and added a new one afterwards. This doesn't work for the input
"/" as it results in an empty string and addsuffix won't add anything
to aan empty string.
Adding a slash to the end first and then removing the possible duplicate
fixes this.

Signed-off-by: Thomas Kühnel <thomas.kuehnel@avm.de>